Added a sed bug workaround and vec2mat workaround
In Window's 10 systems (at least), the GNUWin32 version of sed doesn't remove the temporary files when sed it run with the -i flagtag. I added a workaround. Newer versions of MATLAB (2018a and maybe older) requires the Communications System Toolbox to use the vec2mat function. I replaced that function call with a reshape function call.